+++ lib/cli/commands/clean.js
@@ -1,3 +1,5 @@
+const { readdirSync, removeSync } = require('fs-extra')
+
 const KEEPS = [
   'templates',
   'scripts',
@@ -16,9 +18,7 @@ module.exports = {
   name: 'clean',
   fullName: 'clean',
   description: 'Remove generated files',
-  action(argv, { logger, fs }) {
-    const { readdirSync, removeSync } = fs
-
+  action(argv, { logger }) {
     try {
       readdirSync(process.cwd())
         .filter(file => file[0] !== '.' && KEEPS.indexOf(file) === -1)
