+++ lib/cli/commands/version.js
@@ -1,10 +1,8 @@
-const { version } = require('../../../package.json')
-
 module.exports = {
   name: '--version',
   alias: '-v',
   description: 'Show current version',
-  action(argv, { logger }) {
+  action(argv, { logger, version }) {
     const { version: node, platform } = process
     logger.log(`acyort: ${version}`, `\nnode: ${node}`, `\nos: ${platform}`)
   },
