+++ lib/config/index.js
@@ -28,5 +28,3 @@ module.exports = (base = process.cwd()) => {
     return null
   }
 }
-
-module.exports.defaults = defaults
